move settlement report stuff so subfolder
