+++ source/settlement-report/SettlementReports.Application/SettlementReports/Commands/CancelSettlementReportCommand.cs
@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.SettlementReport.Application.Commands;
+using Energinet.DataHub.SettlementReport.Interfaces.SettlementReports_v2.Models;
 
-namespace Energinet.DataHub.SettlementReport.Application.Handlers;
+namespace Energinet.DataHub.SettlementReport.Application.SettlementReports.Commands;
 
-public interface ICancelSettlementReportJobHandler
-{
-    Task HandleAsync(CancelSettlementReportCommand cancelSettlementReportCommand);
-}
+public sealed record CancelSettlementReportCommand(
+    ReportRequestId RequestId,
+    Guid UserId);
