+++ source/settlement-report/SettlementReports.Application/SettlementReports/Handlers/IRequestSettlementReportJobHandler.cs
@@ -12,10 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.SettlementReport.Application.SettlementReports.Commands;
 using Energinet.DataHub.SettlementReport.Interfaces.SettlementReports_v2.Models;
 
-namespace Energinet.DataHub.SettlementReport.Application.Commands;
+namespace Energinet.DataHub.SettlementReport.Application.SettlementReports.Handlers;
 
-public sealed record CancelSettlementReportCommand(
-    ReportRequestId RequestId,
-    Guid UserId);
+public interface IRequestSettlementReportJobHandler
+{
+    /// <summary>
+    /// Request a settlement report job
+    /// </summary>
+    /// <param name="request">An object containing the parameters of for the report request</param>
+    /// <returns>A JobRunId value representing the run id of the requested settlement report.</returns>
+    Task<JobRunId> HandleAsync(RequestSettlementReportCommand request);
+}
