+++ source/settlement-report/SettlementReports.Application/SettlementReports/Handlers/ISettlementReportJobsDownloadHandler.cs
@@ -12,17 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.SettlementReport.Application.Commands;
 using Energinet.DataHub.SettlementReport.Interfaces.SettlementReports_v2.Models;
 
-namespace Energinet.DataHub.SettlementReport.Application.Handlers;
+namespace Energinet.DataHub.SettlementReport.Application.SettlementReports.Handlers;
 
-public interface IRequestSettlementReportJobHandler
+public interface ISettlementReportJobsDownloadHandler
 {
-    /// <summary>
-    /// Request a settlement report job
-    /// </summary>
-    /// <param name="request">An object containing the parameters of for the report request</param>
-    /// <returns>A JobRunId value representing the run id of the requested settlement report.</returns>
-    Task<JobRunId> HandleAsync(RequestSettlementReportCommand request);
+    Task<Stream> DownloadReportAsync(ReportRequestId requestId, Guid actorId, bool isMultitenancy);
 }
