+++ source/settlement-report/SettlementReports.Infrastructure/Helpers/DatabricksJobsHelper.cs
@@ -14,7 +14,6 @@
 
 using Energinet.DataHub.Core.Databricks.Jobs.Abstractions;
 using Energinet.DataHub.SettlementReport.Application;
-using Energinet.DataHub.SettlementReport.Application.Handlers;
 using Energinet.DataHub.SettlementReport.Infrastructure.SqlStatements.Mappers;
 using Energinet.DataHub.SettlementReport.Interfaces.Helpers;
 using Energinet.DataHub.SettlementReport.Interfaces.Models;
